feat(kiloclaw): refactor volume usage functions and move to instance-display module
