+++ apps/web/src/app/(app)/claw/components/InstanceTab.tsx
@@ -5,6 +5,12 @@ import { Activity, Clock, HardDrive, Loader2, RotateCcw, TimerReset } from 'luci
 import type { AnalyticsEngineResponse, ControllerTelemetryRow } from '@/lib/kiloclaw/disk-usage';
 import type { KiloClawDashboardStatus, GatewayProcessStatusResponse } from '@/lib/kiloclaw/types';
 import { Badge } from '@/components/ui/badge';
+import {
+  formatUptime,
+  formatVolumeUsage,
+  getVolumeBarColor,
+  getVolumeUsagePercent,
+} from '@/lib/kiloclaw/instance-display';
 import { formatTs } from './time';
 import { useClawContext } from './ClawContext';
 
@@ -38,48 +44,12 @@ const GATEWAY_STATE_STYLES: Record<
   },
 };
 
-export function formatUptime(seconds: number): string {
-  const days = Math.floor(seconds / 86400);
-  const hours = Math.floor((seconds % 86400) / 3600);
-  const mins = Math.floor((seconds % 3600) / 60);
-  if (days > 0) return `${days}d ${hours}h ${mins}m`;
-  if (hours > 0) return `${hours}h ${mins}m`;
-  return `${mins}m`;
-}
-
 function formatLastExit(lastExit: NonNullable<GatewayProcessStatusResponse['lastExit']>): string {
   const code = lastExit.code ?? 'null';
   const signal = lastExit.signal ?? 'none';
   const at = new Date(lastExit.at);
   const timeStr = at.toLocaleString();
   return `exit ${code} / ${signal} at ${timeStr}`;
-}
-
-export function formatBytes(bytes: number): string {
-  if (bytes === 0) return '0 B';
-  const units = ['B', 'KB', 'MB', 'GB'];
-  const i = Math.min(Math.floor(Math.log(bytes) / Math.log(1024)), units.length - 1);
-  const value = bytes / 1024 ** i;
-  return `${value.toFixed(i === 0 ? 0 : 1)} ${units[i]}`;
-}
-
-export function formatVolumeUsage(used: number | null, total: number | null): string {
-  if (used === null || total === null) return '—';
-  const raw = (used / total) * 100;
-  const pct = raw % 1 === 0 ? raw.toFixed(0) : (Math.round(raw * 10) / 10).toFixed(1);
-  return `${formatBytes(used)} of ${formatBytes(total)} (${pct}%)`;
-}
-
-export function getVolumeUsagePercent(used: number | null, total: number | null): number | null {
-  if (used === null || total === null || total <= 0) return null;
-  return Math.max(0, Math.min(100, (used / total) * 100));
-}
-
-export function getVolumeBarColor(percent: number | null): string {
-  if (percent === null) return 'bg-emerald-500';
-  if (percent >= 90) return 'bg-red-500';
-  if (percent >= 75) return 'bg-amber-500';
-  return 'bg-emerald-500';
 }
 
 export function diskUsageQueryKey(organizationId: string | undefined) {
